Delete IncidentManagementTeamD2Repository and add functions directly to DiseaseOutbreakEventD2Repository
