+++ src/CompositionRoot.ts
@@ -44,11 +44,8 @@ import { GetTotalCardCountsUseCase } from "./domain/usecases/GetTotalCardCountsU
 import { RoleRepository } from "./domain/repositories/RoleRepository";
 import { RoleD2Repository } from "./data/repositories/RoleD2Repository";
 import { RoleTestRepository } from "./data/repositories/test/RoleTestRepository";
-import { IncidentManagementTeamTestRepository } from "./data/repositories/test/IncidentManagementTeamTestRepository";
-import { IncidentManagementTeamD2Repository } from "./data/repositories/IncidentManagementTeamD2Repository";
-import { IncidentManagementTeamRepository } from "./domain/repositories/IncidentManagementTeamRepository";
-import { GetIncidentManagementTeamByIdUseCase } from "./domain/usecases/GetIncidentManagementTeamByIdUseCase";
 import { DeleteIncidentManagementTeamMemberRoleUseCase } from "./domain/usecases/DeleteIncidentManagementTeamMemberRoleUseCase";
+import { GetIncidentManagementTeamByIdUseCase } from "./domain/usecases/GetIncidentManagementTeamByIdUseCase";
 
 export type CompositionRoot = ReturnType<typeof getCompositionRoot>;
 
@@ -64,7 +61,6 @@ type Repositories = {
     mapConfigRepository: MapConfigRepository;
     performanceOverviewRepository: PerformanceOverviewRepository;
     roleRepository: RoleRepository;
-    incidentManagementTeamRepository: IncidentManagementTeamRepository;
 };
 
 function getCompositionRoot(repositories: Repositories) {
@@ -118,7 +114,6 @@ export function getWebappCompositionRoot(api: D2Api) {
         mapConfigRepository: new MapConfigD2Repository(api),
         performanceOverviewRepository: new PerformanceOverviewD2Repository(api, dataStoreClient),
         roleRepository: new RoleD2Repository(api),
-        incidentManagementTeamRepository: new IncidentManagementTeamD2Repository(api),
     };
 
     return getCompositionRoot(repositories);
@@ -137,7 +132,6 @@ export function getTestCompositionRoot() {
         mapConfigRepository: new MapConfigTestRepository(),
         performanceOverviewRepository: new PerformanceOverviewTestRepository(),
         roleRepository: new RoleTestRepository(),
-        incidentManagementTeamRepository: new IncidentManagementTeamTestRepository(),
     };
 
     return getCompositionRoot(repositories);
