+++ src/data/repositories/test/DiseaseOutbreakEventTestRepository.ts
@@ -5,9 +5,13 @@ import {
     NationalIncidentStatus,
 } from "../../../domain/entities/disease-outbreak-event/DiseaseOutbreakEvent";
 import { Future } from "../../../domain/entities/generic/Future";
+import { IncidentManagementTeam } from "../../../domain/entities/incident-management-team/IncidentManagementTeam";
+import { Role } from "../../../domain/entities/incident-management-team/Role";
+import { TeamMember, TeamRole } from "../../../domain/entities/incident-management-team/TeamMember";
 import { Id, ConfigLabel } from "../../../domain/entities/Ref";
 import { DiseaseOutbreakEventRepository } from "../../../domain/repositories/DiseaseOutbreakEventRepository";
 import { FutureData } from "../../api-futures";
+import { INCIDENT_MANAGER_ROLE } from "../consts/IncidentManagementTeamBuilderConstants";
 
 export class DiseaseOutbreakEventTestRepository implements DiseaseOutbreakEventRepository {
     get(id: Id): FutureData<DiseaseOutbreakEventBaseAttrs> {
@@ -109,13 +113,72 @@ export class DiseaseOutbreakEventTestRepository implements DiseaseOutbreakEventR
             },
         ]);
     }
+
     save(_diseaseOutbreak: DiseaseOutbreakEvent): FutureData<Id> {
         return Future.success("");
     }
+
     delete(_id: Id): FutureData<void> {
         throw new Error("Method not implemented.");
     }
+
     getConfigStrings(): FutureData<ConfigLabel[]> {
         throw new Error("Method not implemented.");
+    }
+
+    getIncidentManagementTeam(
+        _diseaseOutbreakId: Id,
+        _teamMembers: TeamMember[],
+        _roles: Role[]
+    ): FutureData<IncidentManagementTeam> {
+        return Future.success(
+            new IncidentManagementTeam({
+                teamHierarchy: [],
+            })
+        );
+    }
+
+    saveIncidentManagementTeamMemberRole(
+        _teamMemberRole: TeamRole,
+        _incidentManagementTeamMember: TeamMember,
+        _diseaseOutbreakId: Id,
+        _roles: Role[]
+    ): FutureData<void> {
+        return Future.success(undefined);
+    }
+
+    deleteIncidentManagementTeamMemberRole(
+        _teamMemberRole: TeamRole,
+        _incidentManagementTeamMember: TeamMember,
+        _diseaseOutbreakId: Id,
+        _roles: Role[]
+    ): FutureData<void> {
+        return Future.success(undefined);
+    }
+
+    getIncidentManagementTeamMember(
+        username: Id,
+        _diseaseOutbreakId: Id,
+        _roles: Role[]
+    ): FutureData<TeamMember> {
+        const teamMember: TeamMember = new TeamMember({
+            id: username,
+            username: username,
+            name: `Team Member Name ${username}`,
+            email: `email@email.com`,
+            phone: `121-1234`,
+            teamRoles: [
+                {
+                    id: "role",
+                    name: "role",
+                    roleId: INCIDENT_MANAGER_ROLE,
+                    reportsToUsername: "reportsToUsername",
+                },
+            ],
+            status: "Available",
+            photo: new URL("https://www.example.com"),
+            workPosition: "workPosition",
+        });
+        return Future.success(teamMember);
     }
 }
