+++ src/domain/repositories/DiseaseOutbreakEventRepository.ts
@@ -1,5 +1,8 @@
 import { FutureData } from "../../data/api-futures";
 import { DiseaseOutbreakEventBaseAttrs } from "../entities/disease-outbreak-event/DiseaseOutbreakEvent";
+import { IncidentManagementTeam } from "../entities/incident-management-team/IncidentManagementTeam";
+import { Role } from "../entities/incident-management-team/Role";
+import { TeamMember, TeamRole } from "../entities/incident-management-team/TeamMember";
 import { ConfigLabel, Id } from "../entities/Ref";
 
 export interface DiseaseOutbreakEventRepository {
@@ -7,4 +10,26 @@ export interface DiseaseOutbreakEventRepository {
     getAll(): FutureData<DiseaseOutbreakEventBaseAttrs[]>;
     save(diseaseOutbreak: DiseaseOutbreakEventBaseAttrs): FutureData<Id>;
     getConfigStrings(): FutureData<ConfigLabel[]>;
+    getIncidentManagementTeam(
+        diseaseOutbreakId: Id,
+        teamMembers: TeamMember[],
+        roles: Role[]
+    ): FutureData<IncidentManagementTeam>;
+    saveIncidentManagementTeamMemberRole(
+        teamMemberRole: TeamRole,
+        incidentManagementTeamMember: TeamMember,
+        diseaseOutbreakId: Id,
+        roles: Role[]
+    ): FutureData<void>;
+    deleteIncidentManagementTeamMemberRole(
+        teamMemberRole: TeamRole,
+        incidentManagementTeamMember: TeamMember,
+        diseaseOutbreakId: Id,
+        roles: Role[]
+    ): FutureData<void>;
+    getIncidentManagementTeamMember(
+        username: Id,
+        diseaseOutbreakId: Id,
+        roles: Role[]
+    ): FutureData<TeamMember>;
 }
