+++ src/domain/usecases/GetIncidentManagementTeamByIdUseCase.ts
@@ -2,7 +2,7 @@ import { FutureData } from "../../data/api-futures";
 import { Maybe } from "../../utils/ts-utils";
 import { IncidentManagementTeam } from "../entities/incident-management-team/IncidentManagementTeam";
 import { Id } from "../entities/Ref";
-import { IncidentManagementTeamRepository } from "../repositories/IncidentManagementTeamRepository";
+import { DiseaseOutbreakEventRepository } from "../repositories/DiseaseOutbreakEventRepository";
 import { RoleRepository } from "../repositories/RoleRepository";
 import { TeamMemberRepository } from "../repositories/TeamMemberRepository";
 import { getIncidentManagementTeamById } from "./utils/incident-management-team/GetIncidentManagementTeamById";
@@ -12,7 +12,7 @@ export class GetIncidentManagementTeamByIdUseCase {
         private options: {
             roleRepository: RoleRepository;
             teamMemberRepository: TeamMemberRepository;
-            incidentManagementTeamRepository: IncidentManagementTeamRepository;
+            diseaseOutbreakEventRepository: DiseaseOutbreakEventRepository;
         }
     ) {}
